feat(absorb): 26 源批量吸收 + 4 项生产接线 (R-P42/R-P79)
- P1 duration-drift SelfTest (LessWrong 时间感知): DurationDriftMonitor + DurationDriftTest
- P2 precision_gate 检索精度门控 (2608.14036): 池规模感知阈值接入 finalize_search
- P3 RPM rank-then-commit (2608.13940): BayesianExperimentDesign rpm_rank_candidates + rank_then_commit
- P4 MCP RAG Tool Optimizer (pentestagent): recommend_tools_rag + embedding_lite bigram 语义分 (pool>128)
- J-Space 强化延续 (A1-A4) + 先前会话未提交工作 (judge/sandbox/tauri/benches/docs/icons)
- 能力树登记: 3 bud + 1 strengthen; KB 20 节点入库 (absorb-node)

+++ src-tauri/src/commands/kb_cmds.rs
@@ -363,7 +363,7 @@ pub fn kb_geo_points_pack(limit: Option<usize>, source: Option<String>) -> Resul
         .filter(|p| p.exists())
         .unwrap_or(default_path);
 
-    let limit = (limit.unwrap_or(5000) as usize).clamp(1, 20_000);
+    let limit = limit.unwrap_or(5000).clamp(1, 20_000);
     let filtered = |points: Vec<nt_memory_pack::GeoPoint>, limit: usize| {
         points
             .into_iter()
